Moved typings to a .d.ts file
